Made all ulong vars to long

+++ Ceres/Models/FrontMemberInfos.cs
@@ -8,12 +8,12 @@
 
         internal DateTimeOffset EndTime { get; init; }
 
-        public FrontMemberInfos(string memberName, ulong? startTime, ulong? endTime)
+        public FrontMemberInfos(string memberName, long? startTime, long? endTime)
         {
             MemberName = memberName ?? throw new ArgumentNullException(nameof(memberName));
             
-            StartTime = DateTimeOffset.FromUnixTimeMilliseconds((long)(startTime ?? 0L));
-            EndTime = DateTimeOffset.FromUnixTimeMilliseconds((long)(endTime ?? 0L));
+            StartTime = DateTimeOffset.FromUnixTimeMilliseconds(startTime ?? 0L);
+            EndTime = DateTimeOffset.FromUnixTimeMilliseconds(endTime ?? 0L);
         }
     }
 }
